packages/dnsminer/pyproject.toml
[project]
name = "dnsminer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Julien OPPLIGER", email = "account@oppliger.cc" }
]
requires-python = ">=3.13"
dependencies = [
    "open-asset-model>=1.1.1",
    "open-asset-store>=0.0.2",
    "termcolor>=3.3.0",
]

[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project.scripts]
dnsminer = "dnsminer.cli:main"
